243: move responses under json module
