+++ src/rpc_client/json/mod.rs
@@ -1,2 +1,4 @@
+//! Types used for JSON-RPC requests and responses with Ethereum JSON-RPC providers.
+
 pub mod requests;
 pub mod responses;
